# docs/swagger.yaml
basePath: /
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.User:
    properties:
      first_name:
        example: John
        type: string
      id:
        example: 123456789
        type: integer
      language_code:
        example: ru
        type: string
      last_name:
        example: Doe
        type: string
      photo_url:
        example: https://example.com/photo.jpg
        type: string
      username:
        example: johndoe
        type: string
    required:
    - first_name
    - id
    type: object
  github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.UserInfoResponse:
    properties:
      user:
        $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.User'
      user_roles:
        items:
          type: string
        type: array
    type: object
...
  github_com_max-main-team_backend_hackaton_MAX_internal_models_http_subjects.CreateSubjectRequest:
    properties:
      name:
        type: string
      university_id:
        type: integer
    type: object
  github_com_max-main-team_backend_hackaton_MAX_internal_models_http_subjects.SubjectsResponse:
    properties:
      data:
        items:
          properties:
            id:
              type: integer
            name:
              type: string
          type: object
        type: array
      has_more:
        type: boolean
    type: object
...
info:
  contact: {}
  description: API for Hackaton backend
  title: Hackaton API
  version: "1.0"
paths:
...
  /subjects:
    delete:
      consumes:
      - application/json
      description: Delete subject by subject ID. Admin role required.
      parameters:
      - description: Subject ID
        in: query
        name: subject_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Delete subject
      tags:
      - subjects
    get:
      consumes:
      - application/json
      description: Get all subjects for university by university ID with pagination
      parameters:
      - default: 5
        description: Limit of requests, max(50), default(5)
        in: query
        name: limit
        required: true
        type: integer
      - default: 0
        description: Offset, default(0)
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of subjects
          schema:
            $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_models_http_subjects.SubjectsResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Get all subjects by university ID
      tags:
      - subjects
    post:
      consumes:
      - application/json
      description: Create a new subject for the university. Admin role required.
      parameters:
      - description: Subject data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_models_http_subjects.CreateSubjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Create subject for university
      tags:
      - subjects
...
  /user/me:
    get:
      consumes:
      - application/json
      description: Get detailed information about the authenticated user including
        roles
      produces:
      - application/json
      responses:
        "200":
          description: User information with roles
          schema:
            $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.UserInfoResponse'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Get current user information
      tags:
      - user
swagger: "2.0"